relax EnzymeContext constructor to accept any Integer world
Fixes a MethodError on 32-bit platforms where job.world is UInt32.
The struct field remains UInt64 and Julia handles the conversion.

+++ src/Enzyme.jl
@@ -137,7 +137,7 @@ mutable struct EnzymeContext
     modules_to_link::Vector{LLVM.Module}
     edges::Vector{Any}
     nested_cache::Dict{Core.MethodInstance, String}
-    EnzymeContext(world::UInt64) = new(
+    EnzymeContext(world::Integer) = new(
         world,
         LLVM.Module[],
         Any[],
